Market Items component created
